requirements.txt
requests
python-dotenv
langchain
langchain-core
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
